+++ test/config/parse-runtime.js
@@ -68,7 +68,6 @@ describe('parse-runtime', () => {
         expect(config.environment).to.equal('dev');
         expect(config.useDevServer).to.be.true;
         expect(config.devServerUrl).to.equal('http://localhost:8080/');
-        expect(config.useHotModuleReplacement).to.be.false;
         expect(config.devServerKeepPublicPath).to.be.false;
     });
 
@@ -130,14 +129,6 @@ describe('parse-runtime', () => {
         const config = parseArgv(createArgv(['dev']), projectDir);
 
         expect(config.babelRcFileExists).to.be.true;
-    });
-
-    it('dev-server command hot', () => {
-        const testDir = createTestDirectory();
-        const config = parseArgv(createArgv(['dev-server', '--hot']), testDir);
-
-        expect(config.useDevServer).to.be.true;
-        expect(config.useHotModuleReplacement).to.be.true;
     });
 
     it('dev-server command --keep-public-path', () => {